+++ src/Pim/Bundle/CatalogBundle/Doctrine/MongoDBODM/ProductRepository.php
@@ -17,6 +17,7 @@ use Pim\Bundle\CatalogBundle\Entity\Channel;
 use Pim\Bundle\CatalogBundle\Entity\Group;
 use Pim\Bundle\CatalogBundle\Entity\AssociationType;
 use Pim\Bundle\CatalogBundle\Entity\Family;
+use Pim\Bundle\CatalogBundle\Entity\Repository\AttributeRepository;
 use Pim\Bundle\CatalogBundle\Model\AbstractAttribute;
 
 /**
@@ -50,11 +51,8 @@ class ProductRepository extends DocumentRepository implements
      */
     protected $categoryClass;
 
-    /** @var string */
-    protected $identifier;
-
-    /** @var string */
-    protected $attributeClass;
+    /** @var AttributeRepository $attributeRepository */
+    protected $attributeRepository;
 
     /**
      * Set the EntityManager
@@ -81,15 +79,15 @@ class ProductRepository extends DocumentRepository implements
     }
 
     /**
-     * Set the attribute class
+     * Set the attribute repository
      *
-     * @param string $attributeClass
+     * @param AttributeRepository $attributeRepository
      *
      * @return ProductRepository $this
      */
-    public function setAttributeClass($attributeClass)
+    public function setAttributeRepository(AttributeRepository $attributeRepository)
     {
-        $this->attributeClass = $attributeClass;
+        $this->attributeRepository = $attributeRepository;
 
         return $this;
     }
@@ -563,7 +561,7 @@ class ProductRepository extends DocumentRepository implements
         return $this->findOneBy(
             [
                 [
-                    'attribute' => $this->getIdentifier(),
+                    'attribute' => $this->attributeRepository->getIdentifier(),
                     'value' => $code,
                 ]
             ]
@@ -575,33 +573,7 @@ class ProductRepository extends DocumentRepository implements
      */
     public function getReferenceProperties()
     {
-        return array($this->getIdentifier()->getCode());
-    }
-
-    /**
-     * Returns the identifier code
-     *
-     * @return string
-     */
-    public function getIdentifier()
-    {
-        if (!isset($this->identifier)) {
-            if (!$this->entityManager) {
-                throw new \LogicException('Entity Manager must be set before getting reference properties');
-            }
-
-            $this->identifier = $this->entityManager
-                ->createQuery(
-                    sprintf(
-                        'SELECT a FROM %s a WHERE a.attributeType=:identifier_type ',
-                        $this->attributeClass
-                    )
-                )
-                ->setParameter('identifier_type', 'pim_catalog_identifier')
-                ->getSingleResult();
-        }
-
-        return $this->identifier;
+        return array($this->attributeRepository->getIdentifierCode());
     }
 
     /**